refactor: 도메인 객체 수정
