+++ src/main/java/wooteco/subway/path/domain/Distance.java
@@ -4,12 +4,12 @@ import java.util.List;
 import java.util.Objects;
 
 public class Distance {
-    private static final int FIRST_DISTANCE_THRESHOLD = 10;
-    private static final int SECOND_DISTANCE_THRESHOLD = 50;
+    private static final int FIRST_DISTANCE_CHARGE_THRESHOLD = 10;
+    private static final int SECOND_DISTANCE_CHARGE_THRESHOLD = 50;
     private static final int FIRST_CHARGE_DISTANCE = 5;
     private static final int SECOND_CHARGE_DISTANCE = 8;
     private static final int CHARGE_AMOUNT = 100;
-    private static final int FIRST_FULL_CHARGE = 800;
+    private static final int FIRST_DISTANCE_FULL_CHARGE = 800;
 
     private final int distance;
 
@@ -26,23 +26,19 @@ public class Distance {
     }
 
     public boolean isFirstOverCharge() {
-        return distance > FIRST_DISTANCE_THRESHOLD && distance <= SECOND_DISTANCE_THRESHOLD;
+        return distance > FIRST_DISTANCE_CHARGE_THRESHOLD && distance <= SECOND_DISTANCE_CHARGE_THRESHOLD;
     }
 
     public boolean isSecondOverCharge() {
-        return SECOND_DISTANCE_THRESHOLD < distance;
+        return SECOND_DISTANCE_CHARGE_THRESHOLD < distance;
     }
 
     public int calculateFirstOverCharge() {
-        return (int) ((Math.ceil((distance - FIRST_DISTANCE_THRESHOLD) / FIRST_CHARGE_DISTANCE) + 1) * CHARGE_AMOUNT);
+        return (int) ((Math.ceil((distance - FIRST_DISTANCE_CHARGE_THRESHOLD) / FIRST_CHARGE_DISTANCE) + 1) * CHARGE_AMOUNT);
     }
 
     public int calculateSecondOverCharge() {
-        return FIRST_FULL_CHARGE + (int) ((Math.ceil((distance - SECOND_DISTANCE_THRESHOLD) / SECOND_CHARGE_DISTANCE) + 1) * CHARGE_AMOUNT);
-    }
-
-    public int getDistance() {
-        return distance;
+        return FIRST_DISTANCE_FULL_CHARGE + (int) ((Math.ceil((distance - SECOND_DISTANCE_CHARGE_THRESHOLD) / SECOND_CHARGE_DISTANCE) + 1) * CHARGE_AMOUNT);
     }
 
     public Distance add(final Distance that) {
@@ -55,6 +51,10 @@ public class Distance {
 
     public Distance subtract(final Distance that) {
         return new Distance(this.distance - that.distance);
+    }
+
+    public int getDistance() {
+        return distance;
     }
 
     @Override
